feature: change data type

+++ src/main/java/org/sujine/reacttosoundapi/voiceColor/dto/RequestAudioStreamData.java
@@ -6,7 +6,7 @@ import lombok.Getter;
 @AllArgsConstructor
 @Getter
 public class RequestAudioStreamData {
-    byte[] rawStream;
+    double[] rawStream;
     float sampleRate;
     int sampleSize;
     int channel;
